helikon_bot/requirements.txt
anyio==4.11.0
beautifulsoup4==4.14.2
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
exceptiongroup==1.3.0
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
lxml==6.0.2
oauthlib==3.3.1
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.5
pyTelegramBotAPI==4.29.1
python-dotenv==1.1.1
python-telegram-bot==22.5
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.8
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
